+++ app/src/main/res/layout/splash_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.arctouch.codechallenge.splash.SplashActivity">
+    tools:context=".ui.splash.SplashActivity">
 
     <TextView
         android:layout_width="wrap_content"
